// src/interface/noticia.interface.tsx
// Generated by https://quicktype.io

export interface ResultNoticias {
    ok:      boolean;
    msg:     string;
    noticia: Noticia[];
}

export interface Noticia {
    id:          number;
    titulo:      string;
    subtitulo:   string;
    fecha:       string;
    ano:         string;
    descripcion: string;
    estado:      number;
    imagen:      Imagen[];
}

export interface Imagen {
    id:         number;
    nombre:     string;
    id_noticia: number;
}